Safari-611.1.10/LayoutTests/editing/pasteboard/dataTransfer-types-is-initially-empty-on-copy-expected.txt

Tests that DataTransfer's types is initially empty on copy. To manually test, press the button below.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS JSON.stringify(event.clipboardData.types) is "[]"
PASS successfullyParsed is true

TEST COMPLETE
